(feat) Final tweaks for a rendering app bundle. Huzzah!
