+++ client/components/FriendItemList.js
@@ -2,8 +2,12 @@ import {FriendItem} from './FriendItem';
 
 export function FriendItemList({ friends }) {
   return (
-    friends.map(function (friend) {
-      return <FriendItem {...friend} />;
-    })
+    <div className="friendList">
+      {
+        friends.map(function (friend) {
+          return <FriendItem {...friend} />;
+        })
+      }
+    </div>
   );
 };
