+++ client/containers/ProfileContainer.js
@@ -7,6 +7,9 @@ export class ProfileContainer extends React.Component {
 
   constructor ({ user, profileInfoItems, friends }) {
     super();
+    console.log('Profile user:', user);
+    console.log('Profile infos:', profileInfoItems);
+    console.log('Profile friends:', friends);
 
     this.user = user;
     this.profileInfoItems = profileInfoItems;
